pyproject.toml
[tool.poetry]
name = "editor"
version = "1.0.0"
description = ""
authors = []
maintainers = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.101.0"
uvicorn = "^0.23.2"
nltk = "^3.8.1"
kmeans1d = "^0.3.1"
textblob = "^0.17.1"
pyenchant = "^3.2.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
isort = "^5.12.0"
ruff = "^0.0.282"
attrs = "^23.1.0"
httpx = "^0.24.1"
ipdb = "^0.13.13"
pytest-mock = "^3.11.1"

[tool.black]
line-length = 100

[tool.isort]
profile = "black"

[tool.ruff]
line-length = 100
select = ["ALL"]
ignore = [
    "D102",  # Missing docstring in public method
    "D103",  # Missing docstring in public function
    "D104",  # Missing docstring in public package
    "D100",  # Missing docstring in public module
    "D212",  # "multi-line-summary-first-line"
    "D203",  # "one-blank-line-before-class"
    "D107",  # Missing docstring in `__init__`
    "ANN101",  # Missing type annotation for `self` in method
    "D202",  # No blank lines allowed after function docstring (found 1)
    "PT001",  # Use `@pytest.fixture()` over `@pytest.fixture`
]

[tool.ruff.per-file-ignores]
"*/tests/*" = [
    "PT001",  # Use `@pytest.fixture()` over `@pytest.fixture`
]
"test_*" = [
    "S101",  # Use of `assert` detected
    "ANN201",  # Missing return type annotation for public function
]
"entities.py" = [
    "RUF012",  # Mutable class attributes should be annotated with `typing.ClassVar`
]
"entities/*.py" = [
    "RUF012",  # Mutable class attributes should be annotated with `typing.ClassVar`
]
